refactor: rename AlternateHistoryWindow to HistoryWindow and update related imports removed: old historyWindow code
